filter out unautho'ed delete request

+++ src/main/java/com/bumsoap/store/controller/UserCon.java
@@ -122,9 +122,14 @@ public class UserCon {
     @DeleteMapping(UrlMap.DELETE_BY_ID)
     public ResponseEntity<ApiResp> delete(@PathVariable("id") Long id) {
         try {
-            String deletedName = userServ.deleteById(id);
-            return ResponseEntity.ok(
-                    new ApiResp(Feedback.DELETEED_USER_NAME + deletedName, null));
+            if (BsUtils.isQualified(id, false, null)) {
+                String deletedName = userServ.deleteById(id);
+                return ResponseEntity.ok(
+                        new ApiResp(Feedback.DELETEED_USER_NAME + deletedName, null));
+            } else {
+                return ResponseEntity.status(UNAUTHORIZED).body(
+                    new ApiResp(Feedback.NOT_QUALIFIED_FOR + id, null));
+            }
         } catch (Exception e) {
             return ResponseEntity.status(NOT_FOUND)
                     .body(new ApiResp(e.getMessage(), null));
